+++ src/main/kotlin/com/coder/toolbox/util/CoderProtocolHandler.kt
@@ -54,7 +54,10 @@ open class CoderProtocolHandler(
             context.logAndShowInfo("URI will not be handled", "No query parameters were provided")
             return
         }
-
+        // this switches to the main plugin screen, even
+        // if last opened provider was not Coder
+        context.envPageManager.showPluginEnvironmentsPage()
+        markAsBusy()
         if (shouldWaitForAutoLogin) {
             isInitialized.waitForTrue()
         }
@@ -67,12 +70,11 @@ open class CoderProtocolHandler(
         val workspace = restClient.workspaces().matchName(workspaceName, deploymentURL) ?: return
 
         val cli = configureCli(deploymentURL, restClient)
-        reInitialize(restClient, cli)
 
         var agent: WorkspaceAgent
         try {
-            markAsBusy()
-            context.refreshMainPage()
+            reInitialize(restClient, cli)
+            context.envPageManager.showPluginEnvironmentsPage()
             if (!prepareWorkspace(workspace, restClient, workspaceName, deploymentURL)) return
             // we resolve the agent after the workspace is started otherwise we can get misleading
             // errors like: no agent available while workspace is starting or stopping
@@ -86,6 +88,7 @@ open class CoderProtocolHandler(
         } finally {
             unmarkAsBusy()
         }
+        delay(2.seconds)
         val environmentId = "${workspace.name}.${agent.name}"
         context.showEnvironmentPage(environmentId)