+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 ## Unreleased
 
+### Added
+
+- support for basic authentication for HTTP/HTTPS proxy
+- support for Toolbox 2.7 release
+
+### Changed
+
+- improved message while loading the workspace
+
+### Fixed
+
+- URI protocol handler is now able to switch to the Coder provider even if the last opened provider was something else  
+
 ## 0.3.2 - 2025-06-25
 
 ### Changed